requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --no-emit-index-url requirements.in
#
alembic==1.9.4
    # via optuna
blis==0.7.9
    # via thinc
bpemb==0.3.4
    # via embetter
catalogue==2.0.8
    # via
    #   sense2vec
    #   spacy
    #   srsly
    #   thinc
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   nltk
    #   typer
cloudpickle==2.2.1
    # via shap
cmaes==0.9.1
    # via optuna
colorlog==6.7.0
    # via optuna
confection==0.0.4
    # via thinc
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via
    #   matplotlib
    #   yellowbrick
cymem==2.0.7
    # via
    #   preshed
    #   spacy
    #   thinc
embetter[text]==0.3.4
    # via -r requirements.in
feature-engine==1.5.2
    # via -r requirements.in
filelock==3.9.0
    # via
    #   huggingface-hub
    #   transformers
fonttools==4.38.0
    # via matplotlib
fst-pso==1.8.1
    # via pyfume
fuzzytm==2.0.5
    # via gensim
gensim==4.3.0
    # via bpemb
greenlet==2.0.2
    # via sqlalchemy
huggingface-hub==0.12.1
    # via
    #   sentence-transformers
    #   transformers
idna==3.4
    # via requests
importlib-metadata==6.0.0
    # via
    #   alembic
    #   numba
importlib-resources==5.12.0
    # via
    #   alembic
    #   matplotlib
jinja2==3.1.2
    # via spacy
joblib==1.2.0
    # via
    #   nltk
    #   scikit-learn
kiwisolver==1.4.4
    # via matplotlib
langcodes==3.3.0
    # via spacy
llvmlite==0.39.1
    # via numba
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
matplotlib==3.7.0
    # via
    #   -r requirements.in
    #   yellowbrick
miniful==0.0.6
    # via fst-pso
murmurhash==1.0.9
    # via
    #   preshed
    #   spacy
    #   thinc
nltk==3.8.1
    # via sentence-transformers
numba==0.56.4
    # via shap
numpy==1.23.5
    # via
    #   -r requirements.in
    #   blis
    #   bpemb
    #   cmaes
    #   contourpy
    #   feature-engine
    #   fst-pso
    #   fuzzytm
    #   gensim
    #   matplotlib
    #   numba
    #   optuna
    #   pandas
    #   patsy
    #   pyfume
    #   scikit-learn
    #   scipy
    #   sense2vec
    #   sentence-transformers
    #   shap
    #   simpful
    #   spacy
    #   statsmodels
    #   thinc
    #   torchvision
    #   transformers
    #   yellowbrick
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
optuna==3.1.0
    # via -r requirements.in
packaging==23.0
    # via
    #   huggingface-hub
    #   matplotlib
    #   optuna
    #   shap
    #   spacy
    #   statsmodels
    #   thinc
    #   transformers
pandas==1.5.3
    # via
    #   -r requirements.in
    #   embetter
    #   feature-engine
    #   fuzzytm
    #   shap
    #   statsmodels
pathy==0.10.1
    # via spacy
patsy==0.5.3
    # via statsmodels
pillow==9.4.0
    # via
    #   matplotlib
    #   torchvision
plotly==5.13.1
    # via -r requirements.in
preshed==3.0.8
    # via
    #   spacy
    #   thinc
pydantic==1.10.5
    # via
    #   confection
    #   spacy
    #   thinc
pyfume==0.2.25
    # via fuzzytm
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.7.1
    # via pandas
pyyaml==6.0
    # via
    #   huggingface-hub
    #   optuna
    #   transformers
regex==2022.10.31
    # via
    #   nltk
    #   transformers
requests==2.28.2
    # via
    #   bpemb
    #   huggingface-hub
    #   simpful
    #   spacy
    #   torchvision
    #   transformers
scikit-learn==1.1.3
    # via
    #   -r requirements.in
    #   embetter
    #   feature-engine
    #   sentence-transformers
    #   shap
    #   yellowbrick
scipy==1.10.1
    # via
    #   feature-engine
    #   fuzzytm
    #   gensim
    #   pyfume
    #   scikit-learn
    #   sentence-transformers
    #   shap
    #   simpful
    #   statsmodels
    #   yellowbrick
sense2vec==2.0.0
    # via embetter
sentence-transformers==2.2.2
    # via
    #   -r requirements.in
    #   embetter
sentencepiece==0.1.97
    # via
    #   bpemb
    #   sentence-transformers
shap==0.41.0
    # via -r requirements.in
simpful==2.10.0
    # via pyfume
six==1.16.0
    # via
    #   patsy
    #   python-dateutil
slicer==0.0.7
    # via shap
smart-open==6.3.0
    # via
    #   gensim
    #   pathy
    #   spacy
spacy==3.5.0
    # via sense2vec
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.4
    # via spacy
sqlalchemy==2.0.4
    # via
    #   alembic
    #   optuna
srsly==2.4.6
    # via
    #   confection
    #   sense2vec
    #   spacy
    #   thinc
statsmodels==0.13.5
    # via feature-engine
tenacity==8.2.2
    # via plotly
thinc==8.1.7
    # via spacy
threadpoolctl==3.1.0
    # via scikit-learn
tokenizers==0.13.2
    # via transformers
torch==1.13.1
    # via
    #   sentence-transformers
    #   torchvision
torchvision==0.14.1
    # via sentence-transformers
tqdm==4.64.1
    # via
    #   bpemb
    #   huggingface-hub
    #   nltk
    #   optuna
    #   sentence-transformers
    #   shap
    #   spacy
    #   transformers
transformers==4.26.1
    # via
    #   -r requirements.in
    #   sentence-transformers
typer==0.7.0
    # via
    #   pathy
    #   spacy
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   pydantic
    #   sqlalchemy
    #   torch
    #   torchvision
urllib3==1.26.14
    # via requests
wasabi==0.10.1
    # via
    #   sense2vec
    #   spacy
    #   thinc
wheel==0.38.4
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-runtime-cu11
yellowbrick==1.5
    # via -r requirements.in
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
